+++ cmd/mailchain/internal/rest/errs/writer.go
