+++ cmd/mailchain/internal/rest/errs/writer_test.go
